# apps/api/openapi.snapshot.json
{
  "components": {
    "schemas": {
      "AcknowledgementResponse": {
        "additionalProperties": false,
        "properties": {
          "acknowledged_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Acknowledged At"
          },
          "assignment_id": {
            "format": "uuid",
            "title": "Assignment Id",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "recording_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Recording Id"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "id",
          "assignment_id",
          "status",
          "created_at"
        ],
        "title": "AcknowledgementResponse",
        "type": "object"
      },
...
      "AssignmentCreateRequest": {
        "additionalProperties": false,
        "properties": {
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "due_date": {
            "format": "date-time",
            "title": "Due Date",
            "type": "string"
          },
          "pieces": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "maxItems": 10,
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pieces"
          },
          "student_id": {
            "format": "uuid",
            "title": "Student Id",
            "type": "string"
          },
          "studio_id": {
            "format": "uuid",
            "title": "Studio Id",
            "type": "string"
          },
          "techniques": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "maxItems": 15,
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Techniques"
          },
          "title": {
            "maxLength": 500,
            "minLength": 1,
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "studio_id",
          "student_id",
          "title",
          "due_date"
        ],
        "title": "AssignmentCreateRequest",
        "type": "object"
      },
      "AssignmentFeedbackRequest": {
        "additionalProperties": false,
        "properties": {
          "feedback_rating": {
            "maximum": 5.0,
            "minimum": 1.0,
            "title": "Feedback Rating",
            "type": "integer"
          },
          "feedback_text": {
            "maxLength": 5000,
            "minLength": 1,
            "title": "Feedback Text",
            "type": "string"
          }
        },
        "required": [
          "feedback_text",
          "feedback_rating"
        ],
        "title": "AssignmentFeedbackRequest",
        "type": "object"
      },
      "AssignmentResponse": {
        "additionalProperties": false,
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "due_date": {
            "format": "date-time",
            "title": "Due Date",
            "type": "string"
          },
          "feedback_rating": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Feedback Rating"
          },
          "feedback_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Feedback Text"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "pieces": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pieces"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "student_id": {
            "format": "uuid",
            "title": "Student Id",
            "type": "string"
          },
          "studio_id": {
            "format": "uuid",
            "title": "Studio Id",
            "type": "string"
          },
          "teacher_id": {
            "format": "uuid",
            "title": "Teacher Id",
            "type": "string"
          },
          "techniques": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Techniques"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "studio_id",
          "teacher_id",
          "student_id",
          "title",
          "due_date",
          "status",
          "created_at",
          "updated_at"
        ],
        "title": "AssignmentResponse",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "DailyRiff API",
    "version": "0.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/assignments": {
      "get": {
        "operationId": "list_assignments_assignments_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/AssignmentResponse"
                  },
                  "title": "Response List Assignments Assignments Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "description": "Bad request \u2014 invalid input"
          },
          "401": {
            "description": "Not authenticated"
          },
          "405": {
            "description": "Method not allowed"
          },
          "422": {
            "description": "Validation error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Assignments",
        "tags": [
          "assignments"
        ]
      },
      "post": {
        "operationId": "create_assignment_assignments_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssignmentCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssignmentResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "description": "Bad request \u2014 invalid input"
          },
          "401": {
            "description": "Not authenticated"
          },
          "405": {
            "description": "Method not allowed"
          },
          "422": {
            "description": "Validation error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Assignment",
        "tags": [
          "assignments"
        ]
      }
    },
    "/assignments/{assignment_id}": {
      "get": {
        "operationId": "get_assignment_assignments__assignment_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "assignment_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Assignment Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssignmentResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "description": "Bad request \u2014 invalid input"
          },
          "401": {
            "description": "Not authenticated"
          },
          "404": {
            "description": "Assignment not found"
          },
          "405": {
            "description": "Method not allowed"
          },
          "422": {
            "description": "Validation error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Assignment",
        "tags": [
          "assignments"
        ]
      }
    },
    "/assignments/{assignment_id}/acknowledgements": {
      "get": {
        "operationId": "list_acknowledgements_assignments__assignment_id__acknowledgements_get",
        "parameters": [
          {
            "in": "path",
            "name": "assignment_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Assignment Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/AcknowledgementResponse"
                  },
                  "title": "Response List Acknowledgements Assignments  Assignment Id  Acknowledgements Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "description": "Bad request \u2014 invalid input"
          },
          "401": {
            "description": "Not authenticated"
          },
          "404": {
            "description": "Assignment not found"
          },
          "405": {
            "description": "Method not allowed"
          },
          "422": {
            "description": "Validation error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Acknowledgements",
        "tags": [
          "assignments"
        ]
      }
    },
    "/assignments/{assignment_id}/feedback": {
      "post": {
        "operationId": "add_feedback_assignments__assignment_id__feedback_post",
        "parameters": [
          {
            "in": "path",
            "name": "assignment_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Assignment Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssignmentFeedbackRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssignmentResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "description": "Bad request \u2014 invalid input"
          },
          "401": {
            "description": "Not authenticated"
          },
          "404": {
            "description": "Assignment not found"
          },
          "405": {
            "description": "Method not allowed"
          },
          "422": {
            "description": "Validation error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Add Feedback",
        "tags": [
          "assignments"
        ]
      }
    },
...
  }
}